// src/api/back/models/MatchSlotInfo.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Public REST api for dota2classic
 * All stuff
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
import {
    DotaConnectionState,
    DotaConnectionStateFromJSON,
    DotaConnectionStateFromJSONTyped,
    DotaConnectionStateToJSON,
    PlayerInfo,
    PlayerInfoFromJSON,
    PlayerInfoFromJSONTyped,
    PlayerInfoToJSON,
    UserDTO,
    UserDTOFromJSON,
    UserDTOFromJSONTyped,
    UserDTOToJSON,
} from './';

/**
 * 
 * @export
 * @interface MatchSlotInfo
 */
export interface MatchSlotInfo {
    /**
     * 
     * @type {DotaConnectionState}
     * @memberof MatchSlotInfo
     */
    connection: DotaConnectionState;
    /**
     * 
     * @type {UserDTO}
     * @memberof MatchSlotInfo
     */
    user: UserDTO;
    /**
     * 
     * @type {number}
     * @memberof MatchSlotInfo
     */
    team: number;
    /**
     * 
     * @type {PlayerInfo}
     * @memberof MatchSlotInfo
     */
    heroData?: PlayerInfo;
}

export function MatchSlotInfoFromJSON(json: any): MatchSlotInfo {
    return MatchSlotInfoFromJSONTyped(json, false);
}

export function MatchSlotInfoFromJSONTyped(json: any, ignoreDiscriminator: boolean): MatchSlotInfo {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'connection': DotaConnectionStateFromJSON(json['connection']),
        'user': UserDTOFromJSON(json['user']),
        'team': json['team'],
        'heroData': !exists(json, 'heroData') ? undefined : PlayerInfoFromJSON(json['heroData']),
    };
}

export function MatchSlotInfoToJSON(value?: MatchSlotInfo | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'connection': DotaConnectionStateToJSON(value.connection),
        'user': UserDTOToJSON(value.user),
        'team': value.team,
        'heroData': PlayerInfoToJSON(value.heroData),
    };
}